rename sglang_example.py to sglang_rollout.py (#134)
