+++ examples/search-r1/generate_with_search.py
@@ -5,7 +5,7 @@ import re
 from google_search_server import google_search
 from qa_em_format import compute_score_em
 
-from slime.rollout.sglang_example import GenerateState
+from slime.rollout.sglang_rollout import GenerateState
 from slime.utils.http_utils import post
 from slime.utils.types import Sample
