+++ scripts/run-qwen3-235B-A22B-sft.sh
@@ -46,7 +46,7 @@ CKPT_ARGS=(
 )
 
 SFT_ARGS=(
-   --rollout-function-path slime.rollout.sft_example.generate_rollout
+   --rollout-function-path slime.rollout.sft_rollout.generate_rollout
    --prompt-data ${BASE_FOLDER}/openhermes2_5.parquet
    --input-key messages
    --rollout-shuffle
